server/requirements.txt
llama-index
PyMuPDF
python-docx
openpyxl
python-pptx
flask-cors
flask-JWT-Extended
pytest-flask